requirements.txt
python-telegram-bot==21.0.1
SQLAlchemy==2.0.28
python-dotenv==1.0.1